+++ CHANGELOG.md
@@ -10,6 +10,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
     * Add generation of serializers when inheriting models.
 * Add `readonly-cell` helper. It can specify `readonly` property for each cell of list components.
 
+### Fixed
+* `colsconfig-dialog-content` component:
+    * Fix semantic styles for sort direction dropdown.
+
 ## [0.9.0-beta.7] - 2017-04-14
 ### Added
 * Blueprints: